refactor(context-menu): migrate to signals (#418)
