+++ libs/ui/menu/brain/src/lib/brn-menu-trigger.directive.ts
@@ -1,10 +1,34 @@
 import { CdkMenuTrigger } from '@angular/cdk/menu';
-import { Directive } from '@angular/core';
-import { BrnTriggerBase } from './brn-trigger-base';
+import { Directive, effect, inject, input } from '@angular/core';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { BrnMenuAlign, getBrnMenuAlign } from './brn-menu-align';
 
 @Directive({
 	selector: '[brnMenuTriggerFor]',
 	standalone: true,
 	hostDirectives: [{ directive: CdkMenuTrigger, inputs: ['cdkMenuTriggerFor: brnMenuTriggerFor'] }],
 })
-export class BrnMenuTriggerDirective extends BrnTriggerBase {}
+export class BrnMenuTriggerDirective {
+	readonly #cdkTrigger = inject(CdkMenuTrigger, { host: true });
+	public readonly align = input<BrnMenuAlign>(undefined);
+
+	constructor() {
+		// once the trigger opens we wait until the next tick and then grab the last position
+		// used to position the menu. we store this in our trigger which the brnMenu directive has
+		// access to through DI
+		this.#cdkTrigger.opened.pipe(takeUntilDestroyed()).subscribe(() =>
+			setTimeout(
+				() =>
+					// eslint-disable-next-line
+					((this.#cdkTrigger as any)._spartanLastPosition = // eslint-disable-next-line
+						(this.#cdkTrigger as any).overlayRef._positionStrategy._lastPosition),
+			),
+		);
+
+		effect(() => {
+			const align = this.align();
+			if (!align) return;
+			this.#cdkTrigger.menuPosition = getBrnMenuAlign(align);
+		});
+	}
+}
